# docs/openapi.yaml
openapi: 3.1.0
info:
  title: discord-support-hub API
  version: 0.1.0
  summary: API-driven provisioning and governance of isolated private Discord support spaces.
  description: |
    The control surface (FR-11) for discord-support-hub, consumed by the backoffice
    (no human UI in v1). Postgres is the authorization source of truth; Discord is a
    projection the bot reconciles (three-layer truth model).

    Conventions:
      * **Mutating endpoints are asynchronous.** They validate, persist the *desired*
        state, enqueue a job, and return `202 Accepted` with a `job` handle. Poll
        `GET /jobs/{jobId}` for completion. (NFR-2: latency decoupled from Discord rate limits.)
      * **Read endpoints** are served from cache (TTL + write-invalidation) and return `200`.
      * **Authorization always resolves against Postgres**, never the Discord role (NFR-13).
      * **Idempotency:** mutating requests accept an `Idempotency-Key` header; a repeated key
        replays the original response instead of re-enqueueing (NFR-3).
      * **Role-per-merchant access.** Provisioning a space auto-creates a Discord role for the
        merchant and grants that role VIEW+SEND on the channel. A collaborator acquires access by
        joining through the merchant's native invite-with-role link (created once by the operator
        in the Discord client and stored via `PUT /merchants/{merchantId}/invite`), which the hub
...
  license:
    name: Apache-2.0
    identifier: Apache-2.0
...
servers:
...
security:
  - ServiceApiKey: []
...
paths:
...
  /agents:
    get:
      tags: [Agents]
      operationId: listAgents
      summary: List agents (Admin only) (FR-23).
      responses:
        '200':
          description: Agent roster.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Agent' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Agents]
      operationId: addAgent
      summary: Add an agent to the roster (Admin only) (FR-23).
      description: |
        Synchronously records `type=agent` (+ `is_admin`) in Postgres - the authZ source of truth.
        A reconcile job projects the Agent role onto the agent's Discord identity once they are
        present in the guild. Returns `201`.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddAgentRequest'
      responses:
        '201':
          description: Agent recorded.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agent' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /agents/{userId}:
    delete:
      tags: [Agents]
      operationId: removeAgent
      summary: Remove an agent from the roster (Admin only) (FR-23, async).
      description: |
        Marks the agent inactive in Postgres and enqueues a job to remove the Agent role
        (and optionally the guild membership). Returns `202`.
      parameters:
        - $ref: '#/components/parameters/UserId'
        - name: scope
          in: query
          required: false
          description: '`role` removes only the Agent role; `server` also removes from the guild.'
          schema:
            type: string
            enum: [role, server]
            default: role
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '202': { $ref: '#/components/responses/JobAccepted' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ----------------------------------------------------------------------- #
  # TRANSVERSAL
  # ----------------------------------------------------------------------- #
...
components:
  securitySchemes:
    ServiceApiKey:
      type: http
      scheme: bearer
      description: |
        Opaque service API key issued to the backoffice (AuthZ Layer A). Sent as
        `Authorization: Bearer <key>`. Stored hashed at rest; instantly revocable. The future POC
        frontend uses a separate session principal (not this long-lived key) behind the same
        authorization layer.
  parameters:
...
    UserId:
      name: userId
      in: path
      required: true
      schema: { type: string, format: uuid }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Client-supplied key making a mutating request idempotent (NFR-3). A repeated key with the
        same body replays the original response; a repeated key with a different body returns 409.
      schema: { type: string, maxLength: 255 }
...
  headers:
    JobLocation:
      description: URL to poll for job completion.
      schema: { type: string, format: uri }
  responses:
    JobAccepted:
      description: |
        Accepted for asynchronous processing. The desired state is persisted; a worker will apply
        the Discord-side change. Poll the returned job for completion.
      headers:
        Location: { $ref: '#/components/headers/JobLocation' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/JobAcceptedBody' }
    BadRequest:
      description: Invalid request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing or invalid service API key.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Authenticated but not authorized (resolved against Postgres, NFR-13).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Idempotency-Key reused with a different body, or a conflicting state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
...
    JobStatus:
      type: string
      enum: [pending, active, completed, retrying, archived]
...
    AddAgentRequest:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
        display_name: { type: [string, 'null'] }
        discord_user_id:
          type: [string, 'null']
          description: The agent's Discord id, supplied by the admin; the Agent role is projected onto it.
        is_admin:
          type: boolean
          default: false
          description: Grant the Admin roster-management privilege.
...
    Agent:
      type: object
      required: [id, type, is_admin]
      properties:
        id: { type: string, format: uuid }
        type:
          type: string
          enum: [agent]
        is_admin: { type: boolean }
        email: { type: [string, 'null'], format: email }
        display_name: { type: [string, 'null'] }
        discord_user_id: { type: [string, 'null'] }
        provisioned_at:
          type: [string, 'null']
          format: date-time
          description: Set once the bot has added them to the guild and projected the Agent role.
        is_active: { type: boolean }
        created_at: { type: string, format: date-time }
...
    JobAcceptedBody:
      type: object
      required: [job]
      properties:
        job: { $ref: '#/components/schemas/Job' }
    Job:
      type: object
      required: [id, kind, status, created_at]
      properties:
        id: { type: string, format: uuid }
        kind:
          type: string
          description: e.g. provision_space, send_collaborator_invite, expel, project_agent_role, change_lifecycle.
        status: { $ref: '#/components/schemas/JobStatus' }
        space_id: { type: [string, 'null'], format: uuid }
        merchant_id: { type: [string, 'null'], format: uuid }
        user_id: { type: [string, 'null'], format: uuid }
        retry_count: { type: integer }
        error:
          type: [string, 'null']
          description: Last error (redacted) if retrying or archived.
        created_at: { type: string, format: date-time }
        completed_at: { type: [string, 'null'], format: date-time }
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code, e.g. validation_error, forbidden, not_found, idempotency_conflict.
        message: { type: string }
        details:
          type: [object, 'null']
          additionalProperties: true